Rename to Microsoft.AspNetCore.Server.IntegrationTesting Fixes #824
