+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/Common/RuntimeArchitecture.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNetCore.Server.Testing
+namespace Microsoft.AspNetCore.Server.IntegrationTesting
 {
     public enum RuntimeArchitecture
     {
