+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/Deployers/IISDeployer.cs
@@ -6,11 +6,11 @@
 using System;
 using System.Linq;
 using System.Threading;
-using Microsoft.AspNetCore.Server.Testing.Common;
+using Microsoft.AspNetCore.Server.IntegrationTesting.Common;
 using Microsoft.Extensions.Logging;
 using Microsoft.Web.Administration;
 
-namespace Microsoft.AspNetCore.Server.Testing
+namespace Microsoft.AspNetCore.Server.IntegrationTesting
 {
     /// <summary>
     /// Deployer for IIS.
