+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/Deployers/IISExpressDeployer.cs
@@ -6,9 +6,9 @@ using System.Diagnostics;
 using System.IO;
 using System.Threading;
 using Microsoft.Extensions.Logging;
-using Microsoft.AspNetCore.Server.Testing.Common;
+using Microsoft.AspNetCore.Server.IntegrationTesting.Common;
 
-namespace Microsoft.AspNetCore.Server.Testing
+namespace Microsoft.AspNetCore.Server.IntegrationTesting
 {
     /// <summary>
     /// Deployment helper for IISExpress.
