+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/Deployers/NginxDeployer.cs
@@ -5,10 +5,10 @@ using System;
 using System.Diagnostics;
 using System.IO;
 using System.Net.Http;
-using Microsoft.AspNetCore.Server.Testing.Common;
+using Microsoft.AspNetCore.Server.IntegrationTesting.Common;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.AspNetCore.Server.Testing
+namespace Microsoft.AspNetCore.Server.IntegrationTesting
 {
     /// <summary>
     /// Deployer for Kestrel on Nginx.
