+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/Properties/AssemblyInfo.cs
