+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/project.json
