+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/xunit/SkipIfEnvironmentVariableNotEnabled.cs
@@ -4,10 +4,10 @@
 using System;
 using Microsoft.AspNetCore.Testing.xunit;
 
-namespace Microsoft.AspNetCore.Server.Testing
+namespace Microsoft.AspNetCore.Server.IntegrationTesting
 {
     /// <summary>
-    /// Skip test if a given environment variable is not enabled. To enable the test, set environment variable 
+    /// Skip test if a given environment variable is not enabled. To enable the test, set environment variable
     /// to "true" for the test process.
     /// </summary>
     [AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
