+++ src/Microsoft.AspNetCore.Server.IntegrationTesting/xunit/SkipOn32BitOSAttribute.cs
@@ -2,22 +2,23 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Diagnostics;
+using System.IO;
 using Microsoft.AspNetCore.Testing.xunit;
 
-namespace Microsoft.AspNetCore.Server.Testing
+namespace Microsoft.AspNetCore.Server.IntegrationTesting
 {
     /// <summary>
-    /// Skips a test if the runtime used to run the test is CoreClr.
+    /// Skips a 64 bit test if the current Windows OS is 32-bit.
     /// </summary>
     [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    public class SkipIfCurrentRuntimeIsCoreClrAttribute : Attribute, ITestCondition
+    public class SkipOn32BitOSAttribute : Attribute, ITestCondition
     {
         public bool IsMet
         {
             get
             {
-                return !Process.GetCurrentProcess().ProcessName.ToLower().Contains("coreclr");
+                // Directory found only on 64-bit OS.
+                return Directory.Exists(Path.Combine(Environment.GetEnvironmentVariable("SystemRoot"), "SysWOW64"));
             }
         }
 
@@ -25,7 +26,7 @@ namespace Microsoft.AspNetCore.Server.Testing
         {
             get
             {
-                return "Cannot run these test variations using CoreCLR runtime as helpers are not available on CoreCLR.";
+                return "Skipping the x64 test since Windows is 32-bit";
             }
         }
     }
